Handle nested test classes in H2FileBasedDatabaseExtension (#334)
* Handle nested test classes in H2FileBasedDatabaseExtension

* Change H2FileBasedDatabaseExtension#beforeAll so that it only creates
   the test database at the top-level, i.e. not when it encounters
   a nested test class.
* Similarly, update  H2FileBasedDatabaseExtension#beforeAll to only
  delete the test database at the top-level

Fixes #333

* Extract JupiterHelpers utility class

Allows testing of the fallback mechanism used in the code
to detect whether a test class is annotated with Nested. Also
allows sharing this code between our extensions (and we have
others that could use the Nested class check, e.g. possibily the
CuratorTestingServerExtension).

* Simplify...if there is a database, just return

If a database already exists, just return since almost
certainly in a nested test class. Warning about a
condition that we don't expect to ever happen does
not add value. If the database does not exist, then
create it and store it.

* Fix the javadoc that is now incorrect.

* Make the afterAll javadoc more clear (I think)

* Remove unused import

* Minor cleanup

* Fix imports in JupiterHelpers (using IntelliJ)
* Fix grammar errors in JupiterHelpers
* Suppress some SQL warnings in the H2FileBasedDatabaseExtension tests
  (these are IntelliJ-specific suppression values)

+++ src/main/java/org/kiwiproject/test/junit/jupiter/H2FileBasedDatabaseExtension.java
@@ -1,5 +1,9 @@
 package org.kiwiproject.test.junit.jupiter;
 
+import static java.util.Objects.nonNull;
+import static org.kiwiproject.test.junit.jupiter.JupiterHelpers.isTestClassNested;
+import static org.kiwiproject.test.junit.jupiter.JupiterHelpers.testClassNameOrNull;
+
 import lombok.Getter;
 import lombok.experimental.Delegate;
 import lombok.extern.slf4j.Slf4j;
@@ -71,29 +75,39 @@ public class H2FileBasedDatabaseExtension implements BeforeAllCallback, AfterAll
     private H2FileBasedDatabase database;
 
     /**
-     * Creates a new H2 file-based database.
+     * Creates a new H2 file-based database if a database does not exist.
      *
      * @param context extension context
      */
     @Override
     public void beforeAll(ExtensionContext context) {
+        if (nonNull(database)) {
+            LOG.trace("A database already exists (we are probably inside a @Nested test class) so not doing anything");
+            return;
+        }
+
+        LOG.trace("Database does not exist; create it");
         database = H2DatabaseTestHelper.buildH2FileBasedDatabase();
-        LOG.trace("Created database: {}", database);
 
         var namespace = createNamespace();
         context.getStore(namespace).put(DATABASE_KEY, database);
+        LOG.trace("Created and stored database: {}", database);
     }
 
     /**
-     * Deletes the H2 file-based database.
+     * Deletes the H2 file-based database unless exiting a nested test class.
      *
      * @param context extension context
      * @throws Exception if the database directory could not be deleted
      */
     @Override
     public void afterAll(ExtensionContext context) throws Exception {
-        LOG.trace("Deleting database: {}", database);
-        FileUtils.deleteDirectory(database.getDirectory());
+        if (isTestClassNested(context)) {
+            LOG.trace("We're in nested class {}, so NOT deleting the database", testClassNameOrNull(context));
+        } else {
+            LOG.trace("Deleting database: {}", database);
+            FileUtils.deleteDirectory(database.getDirectory());
+        }
     }
 
     /**
